Homework6

+++ src/main/java/com/hillel/andriuta/homework/hw6/main.java
@@ -45,8 +45,8 @@ public class main {
 
         int t1Res = player1T1 + player2T1 + player3T1 + player4T1 + player5T1;
         int t2Res = player1T2 + player2T2 + player3T2 + player4T2 + player5T2;
-        double t1Avg = t1Res / 5;
-        double t2Avg = t2Res / 5;
+        double t1Avg = (double) t1Res / 5;
+        double t2Avg = (double) t2Res / 5;
 
         if (t1Avg > t2Avg){
             System.out.println(t1Name + " wins with " + t1Res + " points");
